Save some space in Anki/collection.media and other improvements
This new version of the addon overlays SVG masks on top of the bitmap files
instead of generating a new bitmap per card side.
This is done with CSS and HTML.
These tricks are needed to save space because Qt 4.7 appears to have a bug
in which SVGs linking to bitmaps aren't displayed correctly.

Besides the above, the following features were implemented:

  - Tags are now written in the Image Occlusion Editor
    instead of the normal card editor.
    Tags don't have autocompletion as they do in Anki's note editor.

  - You can now create a Header and Footer for the cards, in the Image Occlusion Editor.
    The header will appear above the image and the footer will appear bellow the image
    in both question and answer sides.
    The Header will be used as the sort field for the cards.

The HTML documentation is not updated to reflect these changes.

The svg-edit.html was touched but no code was changed.